refactor campaign preview readiness router
