+++ src/app/router_registry.py
@@ -186,6 +186,9 @@ from app.routers.dpm_wave_campaign_launch_package import (
 from app.routers.dpm_wave_campaign_lifecycle import (
     router as dpm_wave_campaign_lifecycle_router,
 )
+from app.routers.dpm_wave_campaign_preview_readiness import (
+    router as dpm_wave_campaign_preview_readiness_router,
+)
 from app.routers.dpm_wave_campaign_readiness import (
     router as dpm_wave_campaign_readiness_router,
 )
@@ -437,6 +440,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_wave_campaign_launch_package_router)
     app.include_router(dpm_wave_campaign_launch_actions_router)
     app.include_router(dpm_wave_campaign_lifecycle_router)
+    app.include_router(dpm_wave_campaign_preview_readiness_router)
     app.include_router(dpm_wave_campaign_readiness_router)
     app.include_router(workbench_router)
     app.include_router(workbench_analytics_router)
